requirements.txt
# WEBXES Tech AI Employee — Python Dependencies
python-dotenv>=1.0.0
schedule>=1.2.0
watchdog>=3.0.0
playwright>=1.40.0
Pillow>=10.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0

# Dashboard API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
watchfiles>=0.21.0
